requirements.txt
python-telegram-bot
google-genai
requests
flask